+++ src/models/thread.ts
@@ -1,21 +1,24 @@
 import { ObjectId } from "mongodb";
 import type { userSex, userRole } from "../types/user";
 
-export default interface Thread {
-    id: number;
-    title: string;
-    op: threadOpType;
-    category: number;
-    c: number;
-    conversation: commentType[];
-    score: number;
-    lastModified: Date;
-    createdAt: Date;
-    slink: string;
-    images: { src: string; cid: number }[];
-    pin?: commentType;
-    _id?: ObjectId;
-}
+export type Thread =
+    | { id: number; removed: true }
+    | {
+          id: number;
+          title: string;
+          op: publicUserType;
+          category: number;
+          count: number;
+          conversation: commentType[];
+          score: number;
+          lastModified: Date;
+          createdAt: Date;
+          slink: string;
+          images: { src: string; cid: number }[];
+          pin?: commentType;
+          _id?: ObjectId;
+          admin?: Admin;
+      };
 
 export type publicUserType = {
     id: number;
@@ -24,43 +27,46 @@ export type publicUserType = {
     sex: userSex;
 };
 
-export type commentType = {
-    /** comment id */
-    id: number;
-    /** if removed all below attributes doesn't exist!!! */
-    removed?: true;
-    /** user id */
-    user: publicUserType;
-    /** html string */
-    comment: string;
-    /** comment converted to text */
-    text: string;
-    /** date string */
-    createdAt: Date;
-    /** shortened link */
-    slink: string;
-    images: string[];
-    /** upvotes */
-    U?: number;
-    /** downvotes */
-    D?: number;
-    /** replies */
-    replies?: number[];
-    /** quote **/
-    quote?: commentType;
-    emotions?: Emotion[];
-};
+export type commentType =
+    /** if removed */
+    | { id: number; removed: true }
+    | {
+          /** comment id */
+          id: number;
+          /** user id */
+          user: publicUserType;
+          /** html string */
+          comment: string;
+          /** comment converted to text */
+          text: string;
+          /** date string */
+          createdAt: Date;
+          /** shortened link */
+          slink: string;
+          images: string[];
+          /** upvotes */
+          U?: number;
+          /** downvotes */
+          D?: number;
+          /** replies */
+          replies?: number[];
+          /** quote **/
+          quote?: commentType;
+          emotions?: Emotion[];
+          admin?: Admin;
+      };
 
 export interface Emotion {
     user: number;
     emotion: string /* must be emoji */;
 }
 
-export type threadOpType = {
-    id: number;
-    name: string;
-    sex: userSex;
-    role: userRole;
-};
+export interface Admin {
+    edits?: [{ admin: AdminUser; reason: string }];
+    replies?: [{ admin: AdminUser; reply: string }];
+}
 
+export type AdminUser = publicUserType & { role: "admin" };
+
+export default Thread;
 export type threadType = Thread;