+++ src/plugins/checkComment.ts
@@ -11,19 +11,21 @@ export default async function (
 
     if (!id || !cid) return;
 
-    const comment = (
-        (await threadCl.findOne(
-            {
-                id,
-                conversation: { $elemMatch: { id: cid } },
-            },
-            { projection: { _id: 0, conversation: { $elemMatch: { id: cid } } } }
-        )) as Thread
-    )?.conversation?.[0];
+    const thread = (await threadCl.findOne(
+        {
+            id,
+            conversation: { $elemMatch: { id: cid } },
+        },
+        { projection: { _id: 0, conversation: { $elemMatch: { id: cid } } } }
+    )) as Thread;
+
+    if ("removed" in thread) return res.code(410).send({ error: "Thread removed." });
+
+    const comment = thread?.conversation?.[0];
 
     if (!comment) return res.code(404).send({ error: "Thread or comment not found." });
 
-    if (comment?.removed) return res.code(410).send({ error: "Comment removed." });
+    if ("removed" in comment) return res.code(410).send({ error: "Comment removed." });
 
     return;
 }