+++ src/plugins/checkMuted.ts
@@ -7,12 +7,14 @@ export default async function checkMuted(req: FastifyRequest, res: FastifyReply)
     const user = verifyUser(req.headers.authorization);
     if (!user) return;
 
-    const muted = (
-        (await usersCl.findOne(
-            { id: user.id },
-            { projection: { _id: 0, muted: 1 } }
-        )) as User
-    ).muted;
+    const muted = Boolean(
+        (
+            (await usersCl.findOne(
+                { id: user.id },
+                { projection: { _id: 0, muted: 1 } }
+            )) as User
+        ).mute
+    );
 
     if (muted)
         return res.code(403).send({ error: "Forbidden. You are muted by an admin." });