+++ src/routes/thread/actions/delete.ts
@@ -1,6 +1,7 @@
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
-import { threadCl } from "../../../common";
+import { removedCl, threadCl } from "../../../common";
+import verifyUser from "../../../lib/auth/verify";
 import regex from "../../../lib/regex";
 import requireAdmin from "../../../plugins/requireAdmin";
 
@@ -13,14 +14,30 @@ export default function (
         id: Type.RegEx(regex.integer),
     });
 
+    const schema = Type.Object({
+        reason: Type.String(),
+    });
+
     fastify.delete(
         "/:id",
-        { schema: { params: paramsSchema }, preHandler: [requireAdmin] },
-        async (req: FastifyRequest<{ Params: Static<typeof paramsSchema> }>, res) => {
+        { schema: { params: paramsSchema, body: schema }, preHandler: [requireAdmin] },
+        async (
+            req: FastifyRequest<{
+                Params: Static<typeof paramsSchema>;
+                Body: Static<typeof schema>;
+            }>,
+            res
+        ) => {
             const id = Number(req.params.id);
+            const { reason } = req.body;
+            const admin = verifyUser(req.headers.authorization);
 
-            if (!(await threadCl.deleteOne({ id })).deletedCount)
-                return res.code(404).send({ error: "Thread not found." });
+            const thread = await threadCl.findOne({ id }, { projection: { _id: 0 } });
+            if (!thread) return res.code(404).send({ error: "Thread not found." });
+
+            await removedCl.insertOne({ thread, type: "thread", admin, reason });
+
+            await threadCl.replaceOne({ id }, { id, removed: true });
 
             return res.send({ success: true });
         }