+++ src/routes/thread/comment/actions/delete.ts
@@ -1,6 +1,7 @@
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
-import { threadCl } from "../../../../common";
+import { removedCl, threadCl } from "../../../../common";
+import verifyUser from "../../../../lib/auth/verify";
 import regex from "../../../../lib/regex";
 import Thread from "../../../../models/thread";
 import requireAdmin from "../../../../plugins/requireAdmin";
@@ -15,28 +16,50 @@ export default function (
         cid: Type.RegEx(regex.integer),
     });
 
+    const schema = Type.Object({
+        reason: Type.String(),
+    });
+
     fastify.delete(
         "/:cid",
-        { schema: { params: paramsSchema }, preHandler: [requireAdmin] },
-        async (req: FastifyRequest<{ Params: Static<typeof paramsSchema> }>, res) => {
+        { schema: { params: paramsSchema, body: schema }, preHandler: [requireAdmin] },
+        async (
+            req: FastifyRequest<{
+                Params: Static<typeof paramsSchema>;
+                Body: Static<typeof schema>;
+            }>,
+            res
+        ) => {
             const id = Number(req.params.id);
             const cid = Number(req.params.cid);
+            const { reason } = req.body;
+            const admin = verifyUser(req.headers.authorization);
 
-            const index = (
-                (await threadCl.findOne(
-                    { id, conversation: { $elemMatch: { id: cid } } },
-                    {
-                        projection: {
-                            _id: 0,
-                            index: { $indexOfArray: ["$conversation.id", cid] },
-                        },
-                    }
-                )) as Thread & { index: number }
-            )?.index;
+            const thread = (await threadCl.findOne(
+                { id, conversation: { $elemMatch: { id: cid } } },
+                {
+                    projection: {
+                        _id: 0,
+                        conversation: { $elemMatch: { id: cid } },
+                        index: { $indexOfArray: ["$conversation.id", cid] },
+                    },
+                }
+            )) as Thread & { index: number };
+
+            if ("removed" in thread) return;
+
+            const index = thread?.index;
 
             // index can be 0
             if (index === undefined || index === -1)
                 return res.code(404).send({ error: "Thread or comment not found." });
+
+            await removedCl.insertOne({
+                comment: thread.conversation[0],
+                thread_id: id,
+                admin,
+                reason,
+            });
 
             await threadCl.updateOne(
                 { id },