+++ src/routes/user/actions/unmute.ts
@@ -1,6 +1,7 @@
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { usersCl } from "../../../common";
+import { agenda } from "../../../lib/agenda";
 import regex from "../../../lib/regex";
 import User from "../../../models/user";
 import requireAdmin from "../../../plugins/requireAdmin";
@@ -24,9 +25,11 @@ export default function (
 
             if (!reqUser) return res.status(404).send({ error: "User not found." });
 
-            if (!reqUser.muted) return res.code(409).send({ error: "User not muted." });
+            if (!reqUser.mute) return res.code(409).send({ error: "User not muted." });
 
-            await usersCl.updateOne({ id }, { $unset: { muted: 1 } });
+            await usersCl.updateOne({ id }, { $unset: { mute: 1 } });
+
+            await agenda.cancel({ name: "unmuteUser", data: { userId: id } });
 
             return res.send({ success: true });
         }