+++ src/routes/users/verify.ts
@@ -1,18 +1,11 @@
-//verify email
-/*Syntax: POST /api/verify
-  {
-    email (email used in sign up): string,
-    code (verification code sent to user's email address): string
-  }
-*/
 import dotenv from "dotenv";
 import { usersCl, verificationCl } from "../../common";
 import hash from "hash.js";
 import { Static, Type } from "@sinclair/typebox";
-import { ajv } from "../../lib/ajv";
 import { createToken } from "../../lib/auth/createtoken";
 import User from "../../models/user";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
+import { agenda } from "../../lib/agenda";
 
 dotenv.config();
 
@@ -31,37 +24,41 @@ export default (
 
     fastify.post(
         "/verify",
+        { schema: { body: schema } },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
-            if (!ajv.validate(schema, req.body))
-                return res.code(400).send({ error: "Bad request." });
+            const { email, code } = req.body;
 
             const verificationData = await verificationCl.findOne({
                 type: "register",
-                email: req.body.email,
-                code: req.body.code,
+                email,
+                code,
             });
 
-            if (verificationData?.code !== req.body.code)
+            if (!verificationData)
                 return res
                     .code(401)
                     .send({ error: "Code incorrect or expired, or email not found." });
+
+            const { name, pwd, sex } = verificationData;
 
             const newUserId =
                 (await usersCl.find().sort({ id: -1 }).limit(1).toArray())[0]?.id + 1 ||
                 1;
 
             const newUser: User = {
-                name: verificationData?.name,
+                name,
                 id: newUserId,
-                email: hash.sha256().update(verificationData?.email).digest("hex"),
-                pwd: verificationData?.pwd,
+                email: hash.sha256().update(email).digest("hex"),
+                pwd,
                 role: "user",
                 createdAt: new Date(),
-                sex: verificationData?.sex,
+                sex,
             };
 
             await usersCl.insertOne(newUser);
-            await verificationCl.deleteOne({ type: "register", email: req.body.email });
+            await verificationCl.deleteOne({ type: "register", email });
+
+            await agenda.cancel({ name: "updateVerificationCode", data: { email } });
 
             const token = createToken(newUser);